+++ x/auth/handler.go
@@ -12,45 +12,12 @@ import (
 func NewHandler(ak AccountKeeper, ck types.CertKeeper) sdk.Handler {
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
 		switch msg := msg.(type) {
-		case types.MsgTriggerVesting:
-			return handleMsgTriggerVesting(ctx, ak, ck, msg)
 		case types.MsgManualVesting:
 			return handleMsgManualVesting(ctx, ak, ck, msg)
 		default:
 			return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "Unrecognized cert Msg type: %v", msg.Type())
 		}
 	}
-}
-
-func handleMsgTriggerVesting(ctx sdk.Context, ak AccountKeeper, ck types.CertKeeper, msg types.MsgTriggerVesting) (*sdk.Result, error) {
-	if !ck.IsCertifier(ctx, msg.Certifier) {
-		return nil, sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "not a valid certifier")
-	}
-
-	acc := ak.GetAccount(ctx, msg.Account)
-	if acc == nil {
-		return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownAddress, "account %s does not exist", msg.Account)
-	}
-	tvacc, ok := acc.(*vesting.TriggeredVestingAccount)
-	if !ok {
-		return nil, sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "account does not appear to be a TriggeredVestingAccount")
-	}
-
-	startTime := ctx.BlockTime().Unix()
-	endTime := startTime
-	for _, p := range tvacc.VestingPeriods {
-		endTime += p.Length
-	}
-	tvacc.BaseVestingAccount.EndTime = endTime
-
-	newacc := vesting.NewTriggeredVestingAccountRaw(
-		tvacc.BaseVestingAccount,
-		startTime,
-		tvacc.VestingPeriods,
-		true,
-	)
-	ak.SetAccount(ctx, newacc)
-	return &sdk.Result{}, nil
 }
 
 func handleMsgManualVesting(ctx sdk.Context, ak AccountKeeper, ck types.CertKeeper, msg types.MsgManualVesting) (*sdk.Result, error) {